refactor: 좌우 스크롤 목록 ui를 분리
